role protected HOC added

+++ packages/backend/src/modules/user/token.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { AuthenticationError } from 'apollo-server-express';
@@ -7,6 +7,7 @@ import * as jwt from '../../common/tool/token.tool';
 import { ITokenDocument } from '../../data/schema/token.schema';
 import ITokens from './interfaces/tokens.interface';
 import TokenExpirationTime from '../../common/enum/token-expiration-time.enum';
+import UserService from './user.service';
 
 @Injectable()
 export default class TokenService {
@@ -15,6 +16,7 @@ export default class TokenService {
   public constructor(
     @InjectModel('Token') private tokenModel: Model<ITokenDocument>,
     private readonly logger: LoggerService,
+    @Inject(forwardRef(() => UserService)) private readonly userService: UserService
   ) { }
 
   public async refreshTokens(token: string, device: string): Promise<ITokens> {
@@ -31,16 +33,23 @@ export default class TokenService {
     return this.issueTokens(refreshToken.user, refreshToken.device);
   }
 
-  public async issueTokens(user: string, device: string): Promise<ITokens> {
+  public async issueTokens(userId: string, device: string): Promise<ITokens> {
     const context = `${this.constructor.name}:${this.issueTokens.name}`;
-    const refreshToken = jwt.sign(user, TokenExpirationTime.MONTH);
-    const accessToken = jwt.sign(user, TokenExpirationTime.DAY);
 
-    this.logger.log('Deleting previous user tokens with device', context, { user, device });
-    await this.tokenModel.deleteMany({ user, device });
+    const user = await this.userService.findById(userId);
 
-    this.logger.log('Creating new refresh token for device', context, { user, device });
-    await this.tokenModel.create({ user, token: refreshToken, device });
+    if (!user) {
+      throw new AuthenticationError('User with such email/password does not exist');
+    }
+
+    const refreshToken = jwt.sign(user.role, userId, TokenExpirationTime.MONTH);
+    const accessToken = jwt.sign(user.role, userId, TokenExpirationTime.DAY);
+
+    this.logger.log('Deleting previous user tokens with device', context, { userId, device });
+    await this.tokenModel.deleteMany({ user: userId, device });
+
+    this.logger.log('Creating new refresh token for device', context, { userId, device });
+    await this.tokenModel.create({ user: userId, token: refreshToken, device });
 
     return { refreshToken, accessToken };
   }
